+++ packages/ddd/src/domain/errors/invalid-vo.error.ts
@@ -1,18 +1,20 @@
-import {
-  DomainError,
-  DomainErrorCode,
-  DomainErrorType,
-  Metadata,
-} from '@/shared';
+import { DomainError, Metadata } from '@/shared';
 
 type Params = {
   value: string;
 };
 
 type Props = Metadata<Params>;
-export class InvalidValueObjectError extends DomainError<Props> {
-  public code = 'DOMAIN.INVALID_VALUE' as DomainErrorCode;
-  public type: DomainErrorType = 'DOMAIN.INVALID_VALUE';
+
+export class InvalidValueObjectError extends DomainError<
+  'DOMAIN.INVALID_VALUE',
+  Props
+> {
+  public readonly code = 'DOMAIN.INVALID_VALUE' as const;
+
+  private constructor(msg: string, meta?: Props) {
+    super(msg, meta);
+  }
 
   public static create(
     msg = 'Provided value object is incorrect',